runfiles,Python: remove library from @bazel_tools
Temporarily remove the Python runfiles library
from @bazel_tools and so prevent releasing it in
Bazel 0.12.0.

Motivation: https://github.com/bazelbuild/bazel/issues/4878

Fixing that bug *after* the python library is in
release Bazel would be harder because users would
need to update their code and BUILD files.

Fixing that bug before the library is released is
a lot easier because nobody depends on it yet, and
we can release the library when it's in the right
place.

Closes #4879.

Change-Id: I8adc219410571f9e2a4c8060f5b3239b957fbd5f
PiperOrigin-RevId: 189744985